Fix structure issues

+++ securenative/models/verify_result.go
@@ -2,6 +2,6 @@ package models
 
 type VerifyResult struct {
 	RiskLevel string
-	Score int32
-	Triggers []string
+	Score     float64
+	Triggers  []string
 }
